Initial bindings with MLeap/Tensorflow.

+++ mleap-runtime/src/test/scala/ml/combust/mleap/runtime/function/UserDefinedFunctionSpec.scala
@@ -17,7 +17,7 @@ class UserDefinedFunctionSpec extends FunSpec {
       val udf2: UserDefinedFunction = (v1: Long, v2: Int) => v1 + v2
       val udf3: UserDefinedFunction = (v1: Boolean, v2: Vector) => "hello": Any
       val udf4: UserDefinedFunction = (v1: Seq[Boolean], v2: Option[Seq[String]], v3: Seq[Double]) => "hello"
-      val udf5: UserDefinedFunction = (v1: Double, v2: Double, v3: Double, v4: Double, v5: String) => 55d
+      val udf5: UserDefinedFunction = (v1: Double, v2: Float, v3: Double, v4: Double, v5: String) => 55d
       val udf0custom: UserDefinedFunction = () => MyCustomObject("hello")
 
       assertUdfForm(udf0, StringType())
@@ -25,7 +25,7 @@ class UserDefinedFunctionSpec extends FunSpec {
       assertUdfForm(udf2, LongType(), LongType(), IntegerType())
       assertUdfForm(udf3, AnyType(), BooleanType(), TensorType.doubleVector())
       assertUdfForm(udf4, StringType(), ListType(BooleanType()), ListType(StringType(), isNullable = true), ListType(DoubleType()))
-      assertUdfForm(udf5, DoubleType(), DoubleType(), DoubleType(), DoubleType(), DoubleType(), StringType())
+      assertUdfForm(udf5, DoubleType(), DoubleType(), FloatType(), DoubleType(), DoubleType(), StringType())
       assertUdfForm(udf0custom, MleapContext.defaultContext.customType[MyCustomObject])
     }
   }
